Update bubble pop animation and add summary notes
The bubble-popping animation has been tweaked so it looks better;
now the center of the bubble pops first, like an actual bubble,
instead of being left with a weird cross in the middle that looks
a bit like a bad copy of a tekken powerhit (or whatever you call it).

Also added some DAY-4 summary notes about overall experience of
making this thing.

This project has the dubious distinction of featuring some of the
worst commit messages I've ever written in my life. It's a liberating
experience.